+++ index.d.ts
@@ -3,10 +3,22 @@ import { Plugin } from "rollup";
 declare const analyzer: (options?: AnalyzerOptions) => Plugin;
 export default analyzer;
 
+export interface Module {
+  id: string;
+  size: number;
+  origSize: number;
+  dependents: string[];
+  percent: number;
+  reduction: number;
+  usedExports: string[];
+  unusedExports: string[];
+}
+
 export interface AnalyzerOptions {
   stdout?: boolean;
   limit?: number;
-  filter?: string | string[] | ((id: string) => boolean);
+  filter?: string | string[] | ((moduleObject: Module) => boolean);
+  filterSummary?: boolean;
   root?: string;
   hideDeps?: boolean;
   showExports?: boolean;
@@ -23,15 +35,4 @@ export interface AnalysisObject {
   bundleReduction: number;
   moduleCount: number;
   modules: Module[];
-}
-
-export interface Module {
-  id: string;
-  size: number;
-  origSize: number;
-  dependents: string[];
-  percent: number;
-  reduction: number;
-  usedExports: string[];
-  unusedExports: string[];
 }